% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_2d_longitude_gridded.R
\name{convert_2d_longitude_gridded}
\alias{convert_2d_longitude_gridded}
\title{Convert spatial object longitude to standard format}
\usage{
convert_2d_longitude_gridded(data.in, write.out = FALSE, output.files = NA)
}
\arguments{
\item{data.in}{character vector, list, or SpatRaster. Single file path, vector of file paths, single SpatRaster, or list of SpatRasters/SpatVectors.}

\item{write.out}{logical. If TRUE, writes a netCDF file (requires output.files). If FALSE, returns a list of spatial objects. Default is FALSE.}

\item{output.files}{character vector. Paths for output files if write.out is TRUE. Default is NA.}
}
\value{
A named list of SpatRaster or SpatVector objects with standardized longitudes.
}
\description{
This function converts the longitudinal coordinates of spatial objects from a 0-360 degree scale to a standard -180 to +180 degree scale. It accepts raw spatial objects or file paths and processes them into a standard list.
}
