% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_longitude.R
\name{convert_longitude}
\alias{convert_longitude}
\title{converts SpatRasters from 0:360 longitude to -180:180}
\usage{
convert_longitude(data)
}
\arguments{
\item{data}{SpatRaster object}
}
\value{
a SpatRaster object
}
\description{
descriptions
}
