% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_daily_coord.R
\name{extract_daily_coord}
\alias{extract_daily_coord}
\title{Matches daily coordinates to gridded data}
\usage{
extract_daily_coord(
  input.dir,
  input.prefix,
  input.type,
  output.dir,
  output.prefix,
  coordinates,
  search.radius = 0,
  var.name,
  statistics
)
}
\arguments{
\item{input.dir}{String. The full input directory for daily input files}

\item{input.prefix}{string. The prefix of the input files}

\item{input.type}{string. The type of input data (e.g., 'daily', 'annual'), assumes data within at daily scale}

\item{output.dir}{string. The full output directory for the gridded data}

\item{output.prefix}{string. The prefix of the output files}

\item{coordinates}{dataframe of lat, lon, and date to be extracted from daily data}

\item{search.radius}{numeric. The number of cell "rings" around the closest match to aggregate over. 0 = closest cell, 1 = 3x3 cells around closest, etc}

\item{var.name}{string. The name of the variable being extracted, used for output file naming}

\item{statistics}{character vector. The statistics to be used for the gridded data. Options are 'mean', 'median', 'min', 'max', 'sd', 'var', 'sum'}
}
\value{
a csv with appended values from coordinates input
}
\description{
descriptions
}
