% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_shp.R
\name{import_shp}
\alias{import_shp}
\title{Imports shape file objects various types and returns a spatVector list for other functions}
\usage{
import_shp(shp.file)
}
\arguments{
\item{shp.file}{character vector, list, or SpatRaster. Single file path, vector of file paths, single SpatRaster, or list of SpatRasters representing the spatial data.}
}
\value{
list of spatVectors
}
\description{
Imports shape file objects various types and returns a spatVector list for other functions
}
