% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_2d_summary_gridded.R
\name{make_2d_summary_gridded}
\alias{make_2d_summary_gridded}
\title{Calculates summary statistics on a 2D spatial grid}
\usage{
make_2d_summary_gridded(
  data.in,
  var.name,
  statistics,
  agg.time,
  file.time = "annual",
  shp.file = NA,
  area.names = NULL,
  tz = NA,
  touches = TRUE,
  output.files = NULL,
  write.out = FALSE
)
}
\arguments{
\item{data.in}{character vector, list, or SpatRaster. Single file path, vector of file paths, single SpatRaster, or list of SpatRasters representing the spatial data to be processed.}

\item{var.name}{character. Variable name you wish to extract and process.}

\item{statistics}{character vector. Which statistic(s) to calculate (e.g., c("mean", "max")).}

\item{agg.time}{character. Time scale to aggregate over (e.g., "days", "months", "years", "season").}

\item{file.time}{character. Time scale of the input files ('daily', 'monthly', 'annual'). Default is 'annual'.}

\item{shp.file}{character, SpatVector, SpatRaster, or NA. Shapefile or raster to mask the input data to. Default is NA.}

\item{area.names}{character vector or NULL. Names of shapefile areas you want to retain. Default is NULL.}

\item{tz}{character or NA. Time zone to convert dates to. No correction if NA. Default is NA.}

\item{touches}{logical. If TRUE, all cells touched by lines or polygons will be masked, not just those on the center point. Default is TRUE.}

\item{output.files}{character vector or NULL. Full output file paths corresponding to each input file if write.out is TRUE. Default is NULL.}

\item{write.out}{logical. If TRUE, writes NetCDF files. If FALSE, returns a list of SpatRasterDatasets. Default is FALSE.}
}
\value{
If write.out is TRUE, writes NetCDF files to disk. If FALSE, returns a named list containing SpatRasterDatasets representing the summarized data.
}
\description{
This function aggregates temporal spatial data (like daily rasters) into broader time periods (e.g., months, seasons) and calculates specified summary statistics. It can optionally mask the input to specific shapefile regions before summarization.
}
