% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_2d_summary_ts.R
\name{make_2d_summary_ts}
\alias{make_2d_summary_ts}
\title{Provides summary statistics of 2d gridded data as time series by area}
\usage{
make_2d_summary_ts(
  data.in,
  file.time,
  output.files,
  shp.file,
  area.names,
  var.name,
  agg.time,
  tz = NA,
  statistics,
  touches = T,
  write.out = F
)
}
\arguments{
\item{data.in}{Either a character vector of full input file names for a list of spatRasters}

\item{file.time}{string. What time scale the input files are on ('daily','monthly','annual')? Assumes all monthly or annual files are on a daily timestep}

\item{output.files}{character vector of full output file names corresponding to each input file}

\item{shp.file}{string. Shape file you wish to crop each input file to}

\item{var.name}{string. Variable name you wish to extract}

\item{agg.time}{character. Time scale to calculate climatology over (days,doy, months,season, or  years)}

\item{tz}{string. Time zone to convert. No correction if NA}

\item{statistics}{character vector. Which statistic to calculate}

\item{touches}{logical. If TRUE, all cells touched by lines or polygons will be masked, not just those on the line render path, or whose center point is within the polygon}

\item{write.out}{logical. If TRUE, will write a netCDF file with output.files. If FALSE will return a list of spatRasters}
}
\value{
a dataframe output variable summarized by timestep for each area.names
}
\description{
descriptions
}
