% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_nc_2d.R
\name{mask_nc_2d}
\alias{mask_nc_2d}
\title{Masks a 2D netCDF using a lower and upper value}
\usage{
mask_nc_2d(
  data.in,
  var.name,
  min.value,
  max.value,
  write.out = FALSE,
  output.files = NULL,
  shp.file = NA,
  binary = FALSE,
  area.names = NA
)
}
\arguments{
\item{data.in}{character vector, list, or SpatRaster. Single file path, vector of file paths, single SpatRaster, or list of SpatRasters.}

\item{var.name}{character. Variable name you wish to extract and process.}

\item{min.value}{numeric. Minimum value of the variable to retain.}

\item{max.value}{numeric. Maximum value of the variable to retain.}

\item{write.out}{logical. If TRUE, writes a netCDF file. If FALSE, returns a list of SpatRasters. Default is FALSE.}

\item{output.files}{character vector. Full output file names corresponding to each input item. Required if write.out is TRUE.}

\item{shp.file}{string, SpatVector, or SpatRaster. Shapefile to crop and mask each input file to. Default is NA.}

\item{binary}{logical. Whether the mask should be binary (1 for inside range, NA for outside) or retain original values. Default is FALSE.}

\item{area.names}{character vector. Specific area names to filter the shapefile by before cropping. Default is NA.}
}
\value{
A named list of SpatRasters if write.out is FALSE. Otherwise, writes NetCDF files and returns nothing.
}
\description{
This function filters 2D gridded spatial data by clamping values within a specified minimum and maximum range. It can optionally convert the valid range into a binary mask and crop the data to a specific spatial area prior to masking.
}
