% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract.R
\name{subtract}
\alias{subtract}
\title{Difference two numbers}
\usage{
subtract(x, y)
}
\arguments{
\item{x}{numeric. First number}

\item{y}{numeric. Second number}
}
\value{
Difference of values
}
\description{
descriptions
}
