% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_2d_gridded_nc.R
\name{summary_2d_gridded_nc}
\alias{summary_2d_gridded_nc}
\title{Provides summary statistics on 2d grid}
\usage{
summary_2d_gridded_nc(
  data.in,
  write.out = F,
  output.files,
  shp.file,
  var.name,
  agg.time,
  statistic,
  area.names
)
}
\value{
netCDF file with same time dimensions as input file
}
\description{
descriptions
}
