% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_2d_ts_nc.R
\name{summary_2d_ts_nc}
\alias{summary_2d_ts_nc}
\title{Provides summary statistics of 2d gridded data as time series by area}
\usage{
summary_2d_ts_nc(
  data.in,
  output.files,
  shp.file,
  area.names,
  var.name,
  agg.time,
  statistic,
  write.out = F
)
}
\value{
a dataframe output variable summarized by timestep for each area.names
}
\description{
descriptions
}
