% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_nc_2d.R
\name{crop_nc_2d}
\alias{crop_nc_2d}
\title{Crops a 2D netCDF based on the extent of a shape file}
\usage{
crop_nc_2d(input.files, write.out = F, output.files, shp.file, var.name)
}
\value{
netCDF file with same time dimensions as input file
}
\description{
descriptions
}
