% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_nc_2d.R
\name{crop_nc_2d}
\alias{crop_nc_2d}
\title{Crops a 2D netCDF based on the extent of a shape file}
\usage{
crop_nc_2d(input.files, write.out = F, output.files, shp.file, var.name)
}
\arguments{
\item{input.files}{character vector of full input file names}

\item{write.out}{logical. If TRUE, will write a netCDF file with output.files. If FALSE will return a list of spatRasters}

\item{output.files}{character vector of full output file names corresponding to each input file}

\item{shp.file}{string. Shape file you wish to crop each input file to}

\item{var.name}{string. Variable name you wish to extract}
}
\value{
netCDF file with same time dimensions as input file
}
\description{
descriptions
}
