% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_2d_anomaly_gridded.R
\name{make_2d_anomaly_gridded}
\alias{make_2d_anomaly_gridded}
\title{Provides a gridded climatology based on a reference dates}
\usage{
make_2d_anomaly_gridded(
  data.in,
  climatology,
  output.files,
  shp.file,
  var.name,
  area.names = NA,
  write.out = F
)
}
\value{
netCDF file with same time dimensions as input file
}
\description{
descriptions
}
