% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_2d_anomaly_ts.R
\name{make_2d_anomaly_ts}
\alias{make_2d_anomaly_ts}
\title{Creates an anomaly timeseries from a climatology}
\usage{
make_2d_anomaly_ts(data.in, climatology, output.files, write.out = F)
}
\value{
a dataframe output variable summarized by timestep for each area.names
}
\description{
descriptions
}
