% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_2d_climatology_gridded.R
\name{make_2d_climatology_gridded}
\alias{make_2d_climatology_gridded}
\title{Provides a gridded climatology based on a reference dates}
\usage{
make_2d_climatology_gridded(
  data.in,
  write.out = F,
  output.files,
  shp.file,
  var.name,
  area.names,
  start.time,
  stop.time,
  agg.time,
  statistic
)
}
\value{
netCDF file with same time dimensions as input file
}
\description{
descriptions
}
