% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_2d_climatology_ts.R
\name{make_2d_climatology_ts}
\alias{make_2d_climatology_ts}
\title{Provides a gridded climatology based on a reference dates}
\usage{
make_2d_climatology_ts(
  data.in,
  write.out = F,
  output.files,
  start.time,
  stop.time,
  statistic
)
}
\value{
netCDF file with same time dimensions as input file
}
\description{
descriptions
}
