% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_2d_deg_day_gridded.R
\name{make_2d_deg_day_gridded_nc}
\alias{make_2d_deg_day_gridded_nc}
\title{Provides a gridded summary of degree-day family statistics}
\usage{
make_2d_deg_day_gridded_nc(
  data.in,
  write.out = F,
  output.files,
  shp.file,
  var.name,
  statistic,
  ref.value,
  type,
  area.names
)
}
\value{
netCDF file with same time dimensions as input file
}
\description{
descriptions
}
