% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_2d_deg_day_gridded.R
\name{make_2d_deg_day_gridded_nc}
\alias{make_2d_deg_day_gridded_nc}
\title{Provides a gridded summary of degree-day family statistics}
\usage{
make_2d_deg_day_gridded_nc(
  data.in,
  write.out = F,
  output.files,
  shp.file,
  var.name,
  statistic,
  ref.value,
  type,
  area.names
)
}
\arguments{
\item{data.in}{Either a character vector of full input file names for a list of spatRasters}

\item{write.out}{logical. If TRUE, will write a netCDF file with output.files. If FALSE will return a list of spatRasters}

\item{output.files}{character vector of full output file names corresponding to each input file}

\item{shp.file}{string. Shape file you wish to crop each input file to}

\item{var.name}{string. Variable name you wish to extract}

\item{statistic}{string. Which statistic to calculate ('dd' for degree days, 'nd' for number of days, 'nd.con' for max consecutive number of days)}

\item{ref.value}{numeric. reference point value for threshold}

\item{type}{string. How to use reference point ('above', 'below', or 'raw')}

\item{area.names}{character vector. Names of shape file areas you want to summarise}
}
\value{
netCDF file with same time dimensions as input file
}
\description{
descriptions
}
