% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_2d_deg_day_ts.R
\name{make_2d_deg_day_ts}
\alias{make_2d_deg_day_ts}
\title{Provides a data frame of degree-day family statistics}
\usage{
make_2d_deg_day_ts(
  data.in,
  output.files,
  shp.file,
  area.names,
  var.name,
  statistic,
  ref.value,
  type,
  write.out = F
)
}
\arguments{
\item{data.in}{Either a character vector of full input file names for a list of spatrasters}

\item{output.files}{character vector of full output file names corresponding to each input file}

\item{shp.file}{string. Shape file you wish to crop each input file to}

\item{var.name}{string. Variable name you wish to extract}

\item{statistic}{string. Which statistic to calculate ('dd' for degree days, 'nd' for number of days)}

\item{ref.value}{numeric. reference point value for threshold}

\item{type}{string. How to use reference point ('above', 'below', or 'raw')}

\item{write.out}{logical. If TRUE, will write a netCDF file with output.files. If FALSE will return a list of spatRasters}
}
\value{
a dataframe output variable summarized by timestep for each area.names
}
\description{
descriptions
}
