% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_2d_deg_day_ts.R
\name{make_2d_deg_day_ts}
\alias{make_2d_deg_day_ts}
\title{Provides a data frame of degree-day family statistics}
\usage{
make_2d_deg_day_ts(
  data.in,
  output.files,
  shp.file,
  area.names,
  var.name,
  statistic,
  ref.value,
  type,
  write.out = F
)
}
\value{
a dataframe output variable summarized by timestep for each area.names
}
\description{
descriptions
}
