% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_2d_summary_gridded.R
\name{make_2d_summary_gridded}
\alias{make_2d_summary_gridded}
\title{Provides summary statistics on 2d grid}
\usage{
make_2d_summary_gridded(
  data.in,
  write.out = F,
  file.time = "annual",
  output.files,
  shp.file,
  var.name,
  agg.time,
  tz = NA,
  statistics,
  touches = T,
  area.names
)
}
\arguments{
\item{data.in}{Either a character vector of full input file names for a list of spatRasters}

\item{write.out}{logical. If TRUE, will write a netCDF file with output.files. If FALSE will return a list of spatRasters}

\item{file.time}{string. What time scale the input files are on ('daily','monthly','annual')? Assumes all monthly or annual files are on a daily timestep}

\item{output.files}{character vector of full output file names corresponding to each input file}

\item{shp.file}{string. Shape file you wish to crop each input file to}

\item{var.name}{string. Variable name you wish to extract}

\item{agg.time}{string. Whether to aggregate over. Passed to terra::tapp (e.g. "days", "months", or "years", "season", etc.)}

\item{tz}{string. Time zone to convert. No correction if NA}

\item{statistics}{character vector. Which statistic to calculate}

\item{touches}{logical. If TRUE, all cells touched by lines or polygons will be masked, not just those on the line render path, or whose center point is within the polygon}

\item{area.names}{character vector. Names of shape file areas you want to summarise}
}
\value{
netCDF file with same time dimensions as input file
}
\description{
descriptions
}
