% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_2d_summary_gridded.R
\name{make_2d_summary_gridded}
\alias{make_2d_summary_gridded}
\title{Provides summary statistics on 2d grid}
\usage{
make_2d_summary_gridded(
  data.in,
  write.out = F,
  output.files,
  shp.file,
  var.name,
  agg.time,
  statistic,
  area.names
)
}
\value{
netCDF file with same time dimensions as input file
}
\description{
descriptions
}
