% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_2d_summary_ts.R
\name{make_2d_summary_ts}
\alias{make_2d_summary_ts}
\title{Calculates summary statistics of 2D gridded data as a time series by area}
\usage{
make_2d_summary_ts(
  data.in,
  var.name,
  statistics,
  agg.time,
  file.time = "annual",
  shp.file = NA,
  area.names = NULL,
  tz = NA,
  touches = TRUE,
  output.files = NULL,
  write.out = FALSE
)
}
\arguments{
\item{data.in}{character vector, list, or SpatRaster. Single file path, vector of file paths, single SpatRaster, or list of SpatRasters representing the spatial data.}

\item{var.name}{character. Variable name you wish to extract.}

\item{statistics}{character vector. Which statistic(s) to calculate (e.g., c("mean", "max")).}

\item{agg.time}{character. Time scale to calculate aggregation over (e.g., "days", "months", "season", "years").}

\item{file.time}{character. What time scale the input files are on ('daily', 'monthly', 'annual'). Default is 'annual'.}

\item{shp.file}{character, SpatVector, SpatRaster, or NA. Shapefile or raster to mask the input data to. Default is NA.}

\item{area.names}{character vector or NULL. Names of the areas in the shapefile to extract. Default is NULL.}

\item{tz}{character or NA. Time zone to convert. No correction if NA. Default is NA.}

\item{touches}{logical. If TRUE, all cells touched by lines or polygons will be masked, not just those on the center point. Default is TRUE.}

\item{output.files}{character vector or NULL. Full output file paths corresponding to each processing group. Default is NULL.}

\item{write.out}{logical. If TRUE, writes RDS files. If FALSE, returns a list of data frames. Default is FALSE.}
}
\value{
If write.out is TRUE, writes RDS files to disk. If FALSE, returns a list of data frames summarized by timestep for each area.
}
\description{
This function extracts spatial raster data across specified shapefile regions and aggregates it temporally to produce timeseries summary statistics. It processes inputs by grouping them (e.g., aggregating daily layers into annual time series) and outputs either a list of summarized data frames or writes RDS files directly.
}
