% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_2d_summary_ts.R
\name{make_2d_summary_ts}
\alias{make_2d_summary_ts}
\title{Provides summary statistics of 2d gridded data as time series by area}
\usage{
make_2d_summary_ts(
  data.in,
  output.files,
  shp.file,
  area.names,
  var.name,
  agg.time,
  statistic,
  write.out = F
)
}
\value{
a dataframe output variable summarized by timestep for each area.names
}
\description{
descriptions
}
