% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_nc_2d.R
\name{mask_nc_2d}
\alias{mask_nc_2d}
\title{Masks a 2D netCDF using a lower and upper value}
\usage{
mask_nc_2d(
  data.in,
  write.out = F,
  output.files,
  shp.file,
  var.name,
  min.value,
  max.value,
  binary = F
)
}
\value{
netCDF file with same time dimensions as input file
}
\description{
descriptions
}
